refactor: rename props and unexport const
